+++ showdb/showdb.c
@@ -155,7 +155,7 @@ static void ln_print_self(const ln_self_t *self)
     printf("\",\n");
 
     //key storage
-    // printf(M_QQ("storage_index") ": " M_QQ("%016" PRIx64) ",\n", self->priv_data.storage_index);
+    printf(M_QQ("storage_index") ": " M_QQ("%016" PRIx64) ",\n", self->priv_data.storage_index);
     // printf(M_QQ("storage_seed") ": \"");
     // ucoin_util_dumpbin(stdout, self->priv_data.storage_seed, UCOIN_SZ_PRIVKEY, false);
     // printf("\",\n");
@@ -283,7 +283,8 @@ static void ln_print_self(const ln_self_t *self)
     ucoin_util_dumptxid(stdout, self->commit_local.txid);
     printf("\",\n");
     printf(INDENT4 M_QQ("htlc_num") ": %" PRIu32 ",\n", self->commit_local.htlc_num);
-    printf(INDENT4 M_QQ("commit_num") ": %" PRIu64 "\n", self->commit_local.commit_num);
+    printf(INDENT4 M_QQ("commit_num") ": %" PRIu64 ",\n", self->commit_local.commit_num);
+    printf(INDENT4 M_QQ("revoke_num") ": %" PRIu64 "\n", self->commit_local.revoke_num);
 
     printf(INDENT3 "},\n");
 
@@ -298,7 +299,8 @@ static void ln_print_self(const ln_self_t *self)
     ucoin_util_dumptxid(stdout, self->commit_remote.txid);
     printf("\",\n");
     printf(INDENT4 M_QQ("htlc_num") ": %" PRIu32 ",\n", self->commit_remote.htlc_num);
-    printf(INDENT4 M_QQ("commit_num") ": %" PRIu64 "\n", self->commit_remote.commit_num);
+    printf(INDENT4 M_QQ("commit_num") ": %" PRIu64 ",\n", self->commit_remote.commit_num);
+    printf(INDENT4 M_QQ("revoke_num") ": %" PRIu64 "\n", self->commit_remote.revoke_num);
     printf(INDENT3 "},\n");
 
     printf(INDENT3 M_QQ("funding_sat") ": %" PRIu64 ",\n", self->funding_sat);
@@ -368,7 +370,7 @@ static void ln_print_announce_short(const uint8_t *pData, uint16_t Len)
                 printf(INDENT3 M_QQ("node") ": \"");
                 ucoin_util_dumpbin(stdout, node_pub, UCOIN_SZ_PUBKEY, false);
                 printf("\",\n");
-                char esc_alias[LN_SZ_ALIAS * 2];
+                char esc_alias[LN_SZ_ALIAS * 2 + 1];
                 escape_json_string(esc_alias, node_alias);
                 printf(INDENT3 M_QQ("alias") ": " M_QQ("%s") ",\n", esc_alias);
                 printf(INDENT3 M_QQ("rgbcolor") ": \"#%02x%02x%02x\",\n", msg.rgbcolor[0], msg.rgbcolor[1], msg.rgbcolor[2]);