+++ ucoin/include/ln.h
@@ -52,8 +52,6 @@ extern "C" {
 #define LN_SZ_ONION_ROUTE               (1366)      ///< サイズ:onion-routing-packet
 #define LN_SZ_ALIAS                     (32)        ///< サイズ:alias長
 #define LN_SZ_NOISE_HEADER              (sizeof(uint16_t) + 16)     ///< サイズ:noiseパケットヘッダ
-#define LN_SZ_GFLEN_MAX                 (4)         ///< サイズ:init.gflen最大
-#define LN_SZ_LFLEN_MAX                 (4)         ///< サイズ:init.lflen最大
 #define LN_SZ_FUNDINGTX_VSIZE           (177)       ///< サイズ:funding_txのvsize(nested in BIP16 P2SH形式)
 #define LN_SZ_ERRMSG                    (256)       ///< サイズ:last error文字列
 
@@ -107,6 +105,14 @@ extern "C" {
 #define LN_UGLY_NORMAL                              ///< payment_hashを保存するタイプ
                                                     ///< コメントアウトするとDB保存しなくなるが、revoked transaction closeから取り戻すために
                                                     ///< 相手のアクションが必要となる
+
+#define LN_INIT_LF_OPT_DATALOSS_REQ     (1 << 0)    ///< option_data_loss_protect
+#define LN_INIT_LF_OPT_DATALOSS_OPT     (1 << 1)    ///< option_data_loss_protect
+#define LN_INIT_LF_ROUTE_SYNC           (1 << 3)    ///< initial_routing_sync
+#define LN_INIT_LF_OPT_UPF_SHDN_REQ     (1 << 4)    ///< option_upfront_shutdown_script
+#define LN_INIT_LF_OPT_UPF_SHDN_OPT     (1 << 5)    ///< option_upfront_shutdown_script
+#define LN_INIT_LF_OPT_GSP_QUERY_REQ    (1 << 6)    ///< gossip_queries
+#define LN_INIT_LF_OPT_GSP_QUERY_OPT    (1 << 7)    ///< gossip_queries
 
 
 /**************************************************************************
@@ -522,6 +528,9 @@ typedef struct {
     uint8_t     *p_channel_id;                      ///< 32: channel-id
     uint64_t    next_local_commitment_number;       ///< 8:  next_local_commitment_number
     uint64_t    next_remote_revocation_number;      ///< 8:  next_remote_revocation_number
+    bool        option_data_loss_protect;           ///< true:your_last_per_commitment_secretとmy_current_per_commitment_pointが有効
+    uint8_t     your_last_per_commitment_secret[UCOIN_SZ_PRIVKEY];      ///< 32: your_last_per_commitment_secret
+    uint8_t     my_current_per_commitment_point[UCOIN_SZ_PUBKEY];       ///< 33: my_current_per_commitment_point
 } ln_channel_reestablish_t;
 
 /// @}
@@ -568,7 +577,7 @@ typedef struct {
  *      - p_dataはMALLOC()で確保するため、呼び出し元がMFREE()で解放すること
  */
 typedef struct {
-    uint8_t     channel_id[LN_SZ_CHANNEL_ID];       ///< 32: channel-id
+    uint8_t     *channel_id;                        ///< 32: channel-id
     uint16_t    len;                                ///< 2: byteslen
     char        *p_data;                            ///< エラー文字列(\0あり)
 } ln_error_t;
@@ -942,7 +951,9 @@ typedef struct {
                                                         // remoteには相手から受信した署名
     uint8_t             txid[UCOIN_SZ_TXID];            ///< txid
     uint16_t            htlc_num;                       ///< commit_tx中のHTLC数
-    uint64_t            commit_num;                     ///< commitment_signed送信後にインクリメントする48bitカウンタ(0～)
+    uint64_t            commit_num;                     ///< commitment_number
+                                                        //      commitment_signed送受信でインクリメントする
+    uint64_t            revoke_num;                     ///< 最後にrevoke_and_ack送信した時のcommitment_number
 } ln_commit_data_t;
 
 
@@ -957,7 +968,10 @@ typedef struct {
 
 
 typedef struct {
-    uint64_t                    storage_index;                  ///< 現在のindex
+    uint64_t                    storage_index;                  ///< 自分のstorage_index
+                                                                //      鍵生成してからデクリメントするため、次に生成する際のindexを指している。
+                                                                //      初期値は0xFFFFFFFFFFFF(48bit)。
+                                                                //      初回のcommit_txは0xFF...FFで作成することになる。
     uint8_t                     storage_seed[LN_SZ_SEED];       ///< ユーザから指定されたseed
 
     uint8_t                     priv[LN_FUNDIDX_MAX][UCOIN_SZ_PRIVKEY];
@@ -974,7 +988,8 @@ struct ln_self_t {
 
     //key storage
     ln_derkey_storage           peer_storage;                   ///< key storage(peer)
-    uint64_t                    peer_storage_index;             ///< 現在のindex(peer)
+    uint64_t                    peer_storage_index;             ///< storage index(peer)
+                                                                //      鍵保存してからデクリメントするため、次に保存する際のindexを指している
 
     //funding
     ln_fundflag_t               fund_flag;                      ///< none/funder/fundee
@@ -1103,6 +1118,13 @@ const uint8_t* ln_get_genesishash(void);
  *
  */
 void ln_set_peer_nodeid(ln_self_t *self, const uint8_t *pNodeId);
+
+
+/** init.localfeatures設定
+ * 未設定の場合はデフォルト値が使用される。
+ * 
+ */
+void ln_set_init_localfeatures(uint8_t lf);
 
 
 /** Channel Establish設定